RD-4332 Just merge the v3.1 secrets.py into the v3 one
